api updates, added client folder with latest changes

+++ client/src/pages/adminManagement/DoctorManagement/index.jsx
@@ -17,13 +17,15 @@ import { useSelector } from "react-redux";
 import { uploadFile } from "../../../ApiCalls/dataUpload";
 
 function AdminManagement() {
-  const roleoptions = ["Doctor", "Medical Staff"].map((role, index) => {
-    return (
-      <option key={index} value={role}>
-        {role}
-      </option>
-    );
-  });
+  const roleoptions = ["Doctor", "Medical Staff", "Dialysis Technician"].map(
+    (role, index) => {
+      return (
+        <option key={index} value={role}>
+          {role}
+        </option>
+      );
+    }
+  );
 
   const practicingAtOptions = practicingAtList.map((pat, index) => {
     return (
@@ -173,7 +175,6 @@ function AdminManagement() {
   const handleSubmit = async () => {
     setErrMsg("");
     setSuccessful("Uploading Data");
-
     if (newDoctor.role == "Doctor" && validateDoctorData(newDoctor)) {
       const photourl = await getFileRes(newDoctor.photo);
       const payload = {
@@ -196,10 +197,10 @@ function AdminManagement() {
         specialities: newDoctor.specialities,
         dailyReadings: newDoctor.dailyReadings,
         dialysisReadings: newDoctor.dialysisReadings,
-        changeby:localStorage.getItem("email"),
-        doctorid:newDoctor.id
+        changeby: localStorage.getItem("email"),
+        doctorid: newDoctor.id,
       };
-
+      console.log("ipdating with,", payload.dailyReadings);
       if (!editMode) {
         const response = await registerDoctor(payload);
         if (response.success) {
@@ -234,19 +235,20 @@ function AdminManagement() {
         email: newDoctor.email,
         experience: newDoctor.yearsOfExperience,
         ref: newDoctor.reference || null,
-        phoneno: newDoctor.phoneNo ,
+        phoneno: newDoctor.phoneNo,
         institute: newDoctor.institute,
         address: newDoctor.address,
         practicingAt: newDoctor.practicingAt,
-        resume: resumeurl.data.objectUrl ||null,
+        resume: resumeurl.data.objectUrl || null,
         photo: photourl?.data.objectUrl,
         description: newDoctor.description,
         email_notification: newDoctor.email_notification,
         can_export: newDoctor.can_export,
         specialities: newDoctor.specialities,
-        changeby:localStorage.getItem("email"),
-        doctorid:newDoctor.id
+        changeby: localStorage.getItem("email"),
+        doctorid: newDoctor.id,
       };
+      console.log("docs payload", payload);
       if (!editMode) {
         const response = await registerDoctor(payload);
         if (response.success) {
@@ -276,15 +278,29 @@ function AdminManagement() {
   };
 
   async function handleDelete(id) {
-    setErrMsg("");
-    setSuccessful("Deleting Data");
-    const response = await deleteDoctor(id);
-    if (response.success) {
-      setSuccessful("Delete Successful!");
-      newDoctorDispatch({ type: "all", payload: {} });
-    } else {
-      setErrMsg("Delete Error! " + response.data);
-      setSuccessful("");
+    try {
+      // Display a confirmation dialog
+      const confirmed = window.confirm(
+        "Are you sure you want to delete this doctor?"
+      );
+
+      if (!confirmed) {
+        return; // If the user cancels, exit the function
+      }
+
+      setErrMsg("");
+      setSuccessful("Deleting Data");
+      const response = await deleteDoctor(id);
+      if (response.success) {
+        setSuccessful("Delete Successful!");
+        newDoctorDispatch({ type: "all", payload: {} });
+      } else {
+        setErrMsg(
+          "Delete Error! (Please delete this doctor from the patients list for all the assigned patients before deleting it permantly!)");
+        setSuccessful("");
+      }
+    } catch (error) {
+      console.error("Error deleting doctor:", error);
     }
   }
 
@@ -660,9 +676,9 @@ function AdminManagement() {
                       onChange={(event) => {
                         console.log(event.target.checked);
                         newDoctorDispatch({
-                          type:"dailyReadingsAlerts",
-                        payload: event.target.checked == true ? "yes" : "no",
-                        })
+                          type: "dailyReadingsAlerts",
+                          payload: event.target.checked == true ? "yes" : "no",
+                        });
                       }}
                     />
                     <label className="ms-2 text-sm font-medium text-gray-500">
@@ -677,9 +693,9 @@ function AdminManagement() {
                       onChange={(event) => {
                         console.log(event.target.checked);
                         newDoctorDispatch({
-                          type:"Dialysis_updates",
-                        payload: event.target.checked == true ? "yes" : "no",
-                        })
+                          type: "Dialysis_updates",
+                          payload: event.target.checked == true ? "yes" : "no",
+                        });
                       }}
                     />
                     <label className="ms-2 text-sm font-medium text-gray-500">
